Improve jj-log parsing logic with custom template.
Right now parsing logic will fail if the jj log command elides versions b/c it inserts "elided versions" into the output graph and/or users modify the log template.

This change uses a custom log template so we can be sure about how to parse it. It uses a single lint and separates fields with non printable character so it's easy to find the record limits with a simple string-split. The PR also inserts the extended description as part of the body so expanding the magit section for a commit will show the full description.